apps/python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==1.26.4
umap-learn==0.5.6
pydantic==2.8.2
